pyproject.toml
[tool.poetry]
name = "catalog"
version = "0.1.0"
description = ""
authors = ["seregaalhimenko <seregaalhimenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
django-treebeard = "^4.7"
psycopg2 = "^2.9.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
